btrfs-progs: receive: process encoded_write commands
Add a new btrfs_send_op and support for both dumping and proper receive
processing which does actual encoded writes.

Encoded writes are only allowed on a file descriptor opened with an extra
flag that allows encoded writes, so we also add support for extra flags
when opening a file for writing (in stubs.h).

Signed-off-by: Boris Burkov <boris@bur.io>