btrfs-progs: mkfs: discard free space
When we've finished creating a new filesystem, walk the free-space tree and
call discard for what we find - this is the approach that we use with FITRIM
in the kernel.

The ranges that are free but not already discarded are the metadata blocks
that mkfs.btrfs creates with its intermediate transactions.

Pull-request: #1021
Signed-off-by: Mark Harmstone <mark@harmstone.com>
Signed-off-by: David Sterba <dsterba@suse.com>